+++ src/cr/nimble/dsp/signals.py
@@ -1,4 +1,4 @@
-# Copyright 2021 CR-Nimble Development Team
+# Copyright 2022-Present CR-Suite Development Team
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,3 +11,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+"""
+Synthetic Signals
+"""
+# pylint: disable=W0611
+
+from cr.nimble._src.dsp.synthetic_signals import (
+    chirp,
+    chirp_centered,
+    pulse,
+    gaussian,
+    gaussian_pulse,
+    decaying_sine_wave,
+    transient_sine_wave
+)