+++ CHANGELOG.md
@@ -9,6 +9,25 @@ All notable changes to this project will be documented in this file.
 
 [Documentation](https://cr-nimble.readthedocs.io/en/latest/)
 
+## [0.2.4] - 2022-08-17
+
+[Documentation](https://cr-nimble.readthedocs.io/en/v0.2.4/)
+
+
+### Added
+
+- Digital signal processing utilities moved from cr-sparse to cr-nimble
+- Moved discrete number related functions from cr-sparse.
+- Some sparse vector and matrix processing functionality moved from cr-sparse.
+
+### Removed
+
+- Unnecessary `__init__.py` files removed.
+
+### Notes
+
+- Jax 0.3.14 compatibility
+- Aligning version numbering across sister projects.
 
 ## [0.1.1] - 2021-12-07
 
@@ -27,6 +46,7 @@ All notable changes to this project will be documented in this file.
 Initial release by refactoring code from `cr-nimble`.
 
 
-[Unreleased]: https://github.com/carnotresearch/cr-nimble/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/carnotresearch/cr-nimble/compare/v0.2.4...HEAD
+[0.2.4]: https://github.com/carnotresearch/cr-nimble/compare/0.1.1...v0.2.4
 [0.1.1]: https://github.com/carnotresearch/cr-nimble/compare/0.1.0...v0.1.1
 [0.1.0]: https://github.com/carnotresearch/cr-nimble/releases/tag/v0.1.0